+++ CHANGELOG.md
@@ -36,6 +36,7 @@ Changelog
 - Final version of Daedalus logo added in the top-bar
 - Receive page design update
 - Correct placeholder text for Ada redemption "Ada amount" input
+- Opt-in mode for sending logs to the remote server
 
 ### Fixes